requirements.txt
flask
torch
torchvision
timm
matplotlib
Pillow
pytorch-grad-cam
numpy
opencv-python
